refactor(breaker): move health.rs tests to sibling file for size guard
health.rs crossed the 1500-line / 50KB source-size guard after the breaker
changes. Move the #[cfg(test)] module into health_tests.rs via #[path] (still a
child module of health, so it reaches private items via super::*). No behavior
change.
